creation of util library and break out functions
In order to (I hope) make the transition to an object oriented version easier, I'm beginning to break out some utility functions into their own library. This is the first stage of that.
