+++ setup.py
@@ -7,7 +7,7 @@
 #
 # written by Jeremy Eglen
 # Created: February 24, 2016
-# Last Modified: March 5, 2019
+# Last Modified: November 12, 2019
 # originally written targeting Python 3.4 and 3.5, some testing on 3.6 and has been lightly tested with Python 2.7
 # working with Python 3.7
 
@@ -21,7 +21,7 @@ with open(path.join(this_directory, 'README.md'), encoding='utf-8') as f:
 
 setup(
     name = "sparki_learning",
-    version = "1.5.1.6",
+    version = "1.5.2.0",
     packages = find_packages(),
 
     # Project uses pyserial for bluetooth, so ensure that package gets
@@ -42,7 +42,7 @@ setup(
     license = "Apache License Version 2.0",
     keywords = "sparki learning myro robot",
     url = "https://github.com/radarjd/sparki_learning",   # project home page
-    download_url = "https://github.com/radarjd/sparki_learning/tarball/v1.5.1.6",
+    download_url = "https://github.com/radarjd/sparki_learning/tarball/v1.5.2.0",
     classifiers = [
                    'Development Status :: 5 - Production/Stable',
                    'Intended Audience :: Education',
